AIv2/backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
mediapipe
numpy
torch
torchvision
python-dotenv
pillow
